Refactor entity/component management and add connection status tags
Changed EntityId type from uint64_t to uint32_t across engine, client, and server. Introduced ConnectedComponentTag and ListeningComponentTag for connection status tracking. Refactored query and component access methods to use more descriptive names (e.g., queryEntityIds, queryEntity). Improved TCP socket receive logic to use std::vector<uint8_t> buffers. Added ReceiveTcpMessageSystem and integrated it into the update loop. Updated logging to use correct format specifiers for EntityId. Removed unused member variables and commented out UDP-related code in client and server for future implementation.

+++ TravelersMultiplayerEngine/Engine/include/TME/engine/networkEcs.hpp
@@ -16,7 +16,7 @@
 
 namespace tme::engine
 {
-	using EntityId = uint64_t;
+	using EntityId = uint32_t;
 
 	struct INetworkComponent;
 	struct INetworkSystem;
