+++ TravelersMultiplayerEngine/Engine/internal/iNetworkSystem.hpp
@@ -5,7 +5,7 @@
 
 namespace tme::engine
 {
-	using EntityId = uint64_t;
+	using EntityId = uint32_t;
 
 	class NetworkEcs;
 	struct INetworkSystem
